% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qnr_categories.R
\name{find_categories}
\alias{find_categories}
\title{Find paradata files}
\usage{
find_categories(dir, file_pattern = "\\\\.xlsx", recurse = TRUE)
}
\arguments{
\item{dir}{Character. Directory to scan for categories files}

\item{file_pattern}{Character. Default value is SuSo's default file name.}

\item{recurse}{Boolean. Default value assumes that \code{dir} is parent diectory
that contains child directories that contain the categories.}
}
\value{
Character vector of paths to categories file(s).
}
\description{
Return path(s) of categories files found target path.
}
