% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_answer_options}
\alias{get_answer_options}
\title{Get answer options}
\usage{
get_answer_options(json_path, categories_dir, varname, to_exclude = NULL)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}

\item{categories_dir}{Character. Path to the \code{Categories} directory
of metadata and/or where the \code{.xlsx} files containing reusable categories
can be found.}

\item{varname}{Variable name. Bare name of variable whose answer options to
extract.}

\item{to_exclude}{Numeric vector. Code of answer options to exclude.}
}
\value{
Named numeric vector.
Values are answer codes. Names are answer labels.
}
\description{
Obtain the answer options for a target variable as a set of name-value pairs
in labelled numeric vector.
}
