% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_attachments}
\alias{get_attachments}
\title{Get metadata on attachments}
\usage{
get_attachments(json_path)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}
}
\value{
Data frame with the following columns:
\itemize{
\item \code{object_type}. Character. Simplified object type. Value: \code{attachment}.
\item \code{type}. Character. Value: \code{"Attachment"}.
\item \code{attachment_id}. Character. GUID for the translation.
\item \code{attachment_name}. Character. User-provided name for the attachment. This is how
Designer fields refer to it.
\item \code{attachment_content_id}. Character. System-provided GUID for files in \code{Attachments}
}
}
\description{
Extract metadata about attachments from the questionnaire JSON
as a data frame.
}
