% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_categories}
\alias{get_categories}
\title{Get metadata on categories.}
\usage{
get_categories(json_path)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}
}
\value{
Data frame with the following columns:
\itemize{
\item \code{object_type}. Character. Simplified object type. Value: \verb{reusable category}.
\item \verb{$type}. Character. May not be present in older JSON files.
\item \code{Id}. Character.
\item \code{Name}. Character.
}
}
\description{
Extract metadata about categories from the questionnaire JSON
as a data frame.
}
