% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_lookup_tables}
\alias{get_lookup_tables}
\title{Get metadata on lookup tables.}
\usage{
get_lookup_tables(json_path)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}
}
\value{
Data frame with the following columns:
\itemize{
\item \code{object_type}. Simple object type. Value: \verb{lookup table}.
\item \code{lookup_table_id}. Character. GUID.
\item \code{lookup_table_name}. Character. Name as it appears in Designer.
\item \code{lookup_table_file_name}. Character. Name of file uploaded in Designer.
}
}
\description{
Extract metadata about translations from the questionnaire JSON
as a data frame.
}
