% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_macros}
\alias{get_macros}
\title{Get metadata for macros.}
\usage{
get_macros(json_path)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}
}
\value{
Data frame with the following columns:
\itemize{
\item \code{object_type}. Character. Type of object. Value: \code{macro}.
\item \code{type}. Character. SuSo-provided object type. Value: \code{Macro}.
May not be present in older JSON versions.
\item \code{macro_id}. Character.
\item \code{macro_name}. Character.
\item \code{macro_description}. Character.
\item \code{macro_content}. Character.
}
}
\description{
Extract metadata about macros from the questionnaire JSON
as a data frame.
}
