% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_ms_answers_as_var_labels}
\alias{get_ms_answers_as_var_labels}
\title{Get mutli-select answers as variable labels}
\usage{
get_ms_answers_as_var_labels(qnr_df, varname, to_exclude = NULL)
}
\arguments{
\item{qnr_df}{Data frame returned from \code{parse_questionnaire}}

\item{varname}{Variable name. Bare name of variable whose answer options to
extract.}

\item{to_exclude}{Numeric vector. Code of answer options to exclude.}
}
\value{
Named character variable.
Names are column names of the form \verb{\{variable\}__\{value\}}.
Values are the answer options.
}
\description{
When exported, multi-select questions become a set of variables whose
name is a combination of the variable name and answer option value in
Designer. This function extracts answer options in convenient form to
either label variables in data or columns in display tables,
where expected specification for labelling is of the form
\verb{\{variable\} = \{variable label\}}.
}
