% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_qnr_metadata}
\alias{get_qnr_metadata}
\title{Get metadata on questionnaire document metadata.}
\usage{
get_qnr_metadata(json_path)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}
}
\value{
List with named elements corresponding the metadata attributes.
}
\description{
Extract metadata about the questionnaire itself from the questionnaire JSON
file as a data frame.
}
