% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_questions_by_section}
\alias{get_questions_by_section}
\title{Get questions with the section to which each belongs.}
\usage{
get_questions_by_section(json_path)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}
}
\value{
Data frame with the following columns
\itemize{
\item \code{title} and \code{section_id} (PublicKey) for sections
\item All the columns from \code{get_questions()} and \code{get_variables()}
}
}
\description{
Extract metadata about questions and variables and their parent section
from the questionnaire JSON as a data frame.
}
\seealso{
get_sections get_questions get_variables
}
