% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_static_texts}
\alias{get_static_texts}
\title{Get metadata on static text.}
\usage{
get_static_texts(json_path)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}
}
\value{
Data frame with the following columns:
\itemize{
\item \code{object_type}. Character. Simplified object type. Value: \code{section}.
\item \code{type}. Character. SuSo-provided type. Value: \code{Group.}
\item \code{public_key}. Character. GUID for object.
\item \code{text}. Character.
\item \code{attachment_name}. Character.
\item \code{condition_expression}. Character.
\item \code{hide_if_disabled}. Logical.
\item \verb{validation_expression_*}. Character. Present if validations used.
\item \verb{validation_message_*}. Character. Present if validations used.
\item \verb{validation_severity_*}. Character. Present if validations used.
}
}
\description{
Extract metadata about sections from the questionnaire JSON
as a data frame.
}
