% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_sub_sections}
\alias{get_sub_sections}
\title{Get metadata for sub-sections}
\usage{
get_sub_sections(json_path)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}
}
\value{
Data frame with the following columns:
\itemize{
\item \code{object_type}. Character. Simplified object type. Value: \code{sub-section}.
\item \code{type}. Character. SuSo provided object type. Value: Group.
\item \code{title}. Character.
\item \code{variable_name}. Character.
\item \code{condition_expression}. Character.
\item \code{public_key}. Character. GUID.
}
}
\description{
Extract metadata about sub-sections from the questionnaire JSON
as a data frame.
}
