% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_validations}
\alias{get_validations}
\title{Get metadata for validations}
\usage{
get_validations(json_path)
}
\arguments{
\item{json_path}{Character. Full path to the Survey Solutions questionnaire
JSON file (e.g., \verb{~/my_proj/path/to/document.json}).}
}
\value{
Data frame with following columns:
\itemize{
\item \code{public_key}. Character. GUID of object. Not available for static text.
\item \code{object_type}. Character. Simplified object type. Value: \code{section}.
\item \code{type}. Character.
\item \code{varname}. Character.
\item \code{text}. Character.
\item \code{expression_number}. Numeric.
\item \code{validation_message}. Character.
\item \code{validation_expression}. Character.
\item \code{severity}. Numeric.
}
}
\description{
Get metadata for validations
}
