% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qnr_categories.R
\name{join_categories}
\alias{join_categories}
\title{Join questionnaire metadata and categories metadata}
\usage{
join_categories(qnr_json_df, categories_df)
}
\arguments{
\item{qnr_json_df}{Data frame. Type produced by \code{parse_questionnaire()}.}

\item{categories_df}{Data frame. Type produced by \code{reshape_categories()}.}
}
\value{
Data frame. Same structure as \code{qnr_json_df}, but with categories
joined/updated by \code{categories_id} key.
}
\description{
Join questionnaire metadata and categories metadata
}
