% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qnr_categories.R
\name{parse_categories_file}
\alias{parse_categories_file}
\title{Parse a categories file}
\usage{
parse_categories_file(path = path, sheet = "Categories")
}
\arguments{
\item{path}{Character. Path to a categories file.}

\item{sheet}{Character. Name of the Excel sheet where categories data are
stored. By default, this is Survey Solutions' storage: \code{"Categories"}.}
}
\value{
Data frame with the following columns:
\itemize{
\item \code{category_id}. GUID used in questionnaire JSON reference a reusable
category.
\item \code{value}. Numerical value input in Designer (e.g., \code{1}, \code{2}).
\item \code{text}. Character label input in Designer (e.g., \code{"Yes"}, \code{"No"})
\item \code{parentid}. GUID used to identify parent question.
}
}
\description{
Survey Solutions stores reusable categories in Excel files.
This function parses that file and returns a data frame.
}
