% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qnr_json.R
\name{parse_questionnaire}
\alias{parse_questionnaire}
\title{Parses JSON questionnaire file into rectangular data frame}
\usage{
parse_questionnaire(path)
}
\arguments{
\item{path}{Complete file path to the questionnaire JSON file, typically named \code{document.json}}
}
\value{
Data frame that contains all attributes of all objects in the questionnaire.
}
\description{
Transforms the information found in the JSON questionnaire file, \code{document.json}, into a simpler rectangular file.
The JSON file provides a (deeply) nested representation of the questionnaire. This function returns a rectangular data frame in the following format:
}
\details{
\itemize{
\item Rows: objects in the questionnaire (e.g., sections, sub-sections, rosters, questions, variables, static text, etc.)
\item Columns: attributes.
}

This format facilitates common needs such as:
\itemize{
\item Collecting all the validations
\item Compiling keys and values for questions (e.g., PublicKey for questions to understand the paradata AnswerRemove event)
}
}
