% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qnr_categories.R
\name{reshape_categories}
\alias{reshape_categories}
\title{Reshape categories data}
\usage{
reshape_categories(categories_df)
}
\arguments{
\item{categories_df}{Data frame. categories data frame produced by
\code{parse_categories()} or \code{parse_categories_file()}.}
}
\value{
Data frame with columns:
\itemize{
\item \code{category_id}. Category ID for joining with questionnaire metadata.
\item \verb{answer_value_*}. Value of answer option.
\item \verb{answer_text_*}. Text label of answer option.
}
}
\description{
Reshape categories data to be joined with metadata by \code{category_id}.
}
