% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_answer_options}
\alias{get_answer_options}
\title{Get answer options}
\usage{
get_answer_options(qnr_df, varname, to_exclude = NULL)
}
\arguments{
\item{qnr_df}{Data frame returned from \code{parse_questionnaire}}

\item{varname}{Variable name. Bare name of variable whose answer options to extract.}

\item{to_exclude}{Numeric vector. Code of answer options to exclude.}
}
\value{
Named numeric vector. Values are answer codes. Names are answer labels.
}
\description{
Obtain the answer options for a target variable as a set of name-value pairs in labelled numeric vector.
}
