% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_questions}
\alias{get_questions}
\title{Get questons from questionnaire metadata}
\usage{
get_questions(qnr_df)
}
\arguments{
\item{qnr_df}{Data frame produced by \code{susopara::parse_questionnaire()}}
}
\value{
Data frame of questions (\code{varname}), their JSON indices (\verb{l_*}),
and other question attributes
}
\description{
Get questons from questionnaire metadata
}
