% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_questions_by_section}
\alias{get_questions_by_section}
\title{Get variables by section}
\usage{
get_questions_by_section(qnr_df)
}
\arguments{
\item{qnr_df}{Data frame produced by \code{susopara::parse_questionnaire()}}
}
\value{
Data frame of the mapping of variables to sections
}
\description{
Get variables by section
}
