% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_rosters}
\alias{get_rosters}
\title{Get rosters}
\usage{
get_rosters(qnr_df)
}
\arguments{
\item{qnr_df}{Data frame produced by \code{susopara::parse_questionnaire()}}
}
\value{
Data frame of roster objects, their JSON indices,
and attributes
}
\description{
Get rosters
}
