% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_sections}
\alias{get_sections}
\title{Get sections from questionnaire metadata}
\usage{
get_sections(qnr_df)
}
\arguments{
\item{qnr_df}{Data frame produced by \code{susopara::parse_questionnaire()}}
}
\value{
Data frame of sections (\code{title}), (JSON) indices (\code{l_0}),
and other section attributes
}
\description{
Get sections from questionnaire metadata
}
