% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_validations}
\alias{get_validations}
\title{Get validations from questionnaire metadata}
\usage{
get_validations(qnr_df)
}
\arguments{
\item{qnr_df}{Data frame produced by \code{susopara::parse_questionnaire()}}
}
\value{
Data frame of validated objects and their validation. Objects with
validations include questions and static text. The validation information
returned in the data frame is:
\itemize{
\item \code{type}. Question type drawn directly from the JSON file.
\item \code{varname}. Variable name. Present only for questions.
\item \code{text}. Question text or static text. For static text objects, this
may serve as  practical identifier, for lack of the equivalent of a
unique variable name.
\item \code{expression_number}. Sequential ID number of the validation defined by
Designer.
\item \code{validation_expression}. Expression provided in the \verb{validation condition}
field in Designer.
\item \code{expression_message}. Text composed in the \verb{error or warning message}
field in Designer
\item \code{severity}. Severity level of the validation. Values of 0 denote an error
while values of 1 denote a warning.
}
}
\description{
Get validations from questionnaire metadata
}
